% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.env.functions.R
\name{linearFun}
\alias{linearFun}
\title{Linear function}
\usage{
linearFun(x, a, b)
}
\arguments{
\item{x}{a numeric value or vector}

\item{a}{a numeric value or vector}

\item{b}{a numeric value or vector}
}
\value{
a numeric value or vector resulting from the function
}
\description{
A simple linear function of the form
\deqn{ax+b}{a*x+b}
}
\examples{
x <- 1:100
y <- linearFun(x, a = 0.5, b = 0)
plot(y ~ x, type = "l")
}
\seealso{
\code{\link{logisticFun}}, \code{\link{quadraticFun}}
}
\author{
Boris Leroy \email{leroy.boris@gmail.com}

Maintainer: Boris Leroy \email{leroy.boris@gmail.com}
}
