% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.env.functions.R
\name{quadraticFun}
\alias{quadraticFun}
\title{Quadratic function}
\usage{
quadraticFun(x, a, b, c)
}
\arguments{
\item{x}{a numeric value or vector}

\item{a}{a numeric value or vector}

\item{b}{a numeric value or vector}

\item{c}{a numeric value or vector}
}
\value{
a numeric value or vector resulting from the function
}
\description{
A simple quadratic function of the form
\deqn{ax^2+bx+c}{
a*x^2+b*x+c}
}
\examples{
x <- 1:100
y <- quadraticFun(x, a = 2, b = 2, c = 3)
plot(y ~ x, type = "l")
}
\seealso{
\code{\link{linearFun}}, \code{\link{quadraticFun}}
}
\author{
Boris Leroy \email{leroy.boris@gmail.com}

Maintainer: Boris Leroy \email{leroy.boris@gmail.com}
}
